requirements.txt
pytube~=15.0.0
speedtest-cli~=2.1.3